# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface StageMessages {
  reveal: ChatMessage[];
  identify: ChatMessage[];
  pinpoint: ChatMessage[];
  personalize: ChatMessage[];
  live: ChatMessage[];
  expand: ChatMessage[];
}

export type JourneyRecordStatus =
  (typeof JourneyRecordStatus)[keyof typeof JourneyRecordStatus];

export const JourneyRecordStatus = {
  in_progress: "in_progress",
  complete: "complete",
} as const;

export interface Season {
  summary?: string;
  roles?: string[];
  capacity?: string;
  constraints?: string[];
  opportunities?: string[];
  expressions?: string[];
}

export interface ActionPlan {
  start?: string[];
  stop?: string[];
  continue?: string[];
  oneStepThisWeek?: string;
}

export interface JourneyRecord {
  id: number;
  userId: string;
  status: JourneyRecordStatus;
  /**
   * @minimum 0
   * @maximum 5
   */
  currentStageIdx: number;
  selectedPrompts: string[];
  purposeOptions: string[];
  /** @nullable */
  purposeStatement?: string | null;
  season?: Season;
  actionPlan?: ActionPlan;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface StoryCard {
  id?: number;
  journeyId?: number;
  position?: number;
  title?: string;
  summary?: string;
  /** @nullable */
  promptId?: string | null;
  isDifficultExperience?: boolean;
  actions?: string[];
  feelings?: string[];
  people?: string[];
  impact?: string[];
  userEdited?: boolean;
  createdAt?: string;
  updatedAt?: string;
}

export interface Evidence {
  storyTitle: string;
  detail: string;
}

export interface PurposeTheme {
  id?: number;
  journeyId?: number;
  position?: number;
  name?: string;
  description?: string;
  evidence?: Evidence[];
  userEdited?: boolean;
  createdAt?: string;
  updatedAt?: string;
}

export type Journey = JourneyRecord & {
  messages: StageMessages;
  storyCards: StoryCard[];
  themes: PurposeTheme[];
};

export type JourneyUpdateStatus =
  (typeof JourneyUpdateStatus)[keyof typeof JourneyUpdateStatus];

export const JourneyUpdateStatus = {
  in_progress: "in_progress",
  complete: "complete",
} as const;

export interface JourneyUpdate {
  /**
   * @minimum 0
   * @maximum 5
   */
  currentStageIdx?: number;
  selectedPrompts?: string[];
  purposeOptions?: string[];
  /** @nullable */
  purposeStatement?: string | null;
  status?: JourneyUpdateStatus;
  season?: Season;
  actionPlan?: ActionPlan;
}

export interface StoryCardUpdate {
  title?: string;
  summary?: string;
  /** @nullable */
  promptId?: string | null;
  isDifficultExperience?: boolean;
  actions?: string[];
  feelings?: string[];
  people?: string[];
  impact?: string[];
}

export interface PurposeThemeCreate {
  position?: number;
  name: string;
  description: string;
  evidence?: Evidence[];
}

export type PurposeThemeUpdate = PurposeThemeCreate;

export type ExtractionResponse =
  | {
      stories?: StoryCard[];
    }
  | {
      themes?: PurposeTheme[];
    }
  | {
      season?: Season;
    }
  | {
      actionPlan?: ActionPlan;
    };