# lib/api-zod/src/generated/types/extractionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionPlan } from "./actionPlan";
import type { PurposeTheme } from "./purposeTheme";
import type { Season } from "./season";
import type { StoryCard } from "./storyCard";

export type ExtractionResponse =
  | {
      stories?: StoryCard[];
    }
  | {
      themes?: PurposeTheme[];
    }
  | {
      season?: Season;
    }
  | {
      actionPlan?: ActionPlan;
    };