# lib/api-zod/src/generated/types/journeyRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionPlan } from "./actionPlan";
import type { JourneyRecordStatus } from "./journeyRecordStatus";
import type { Season } from "./season";

export interface JourneyRecord {
  id: number;
  userId: string;
  status: JourneyRecordStatus;
  /**
   * @minimum 0
   * @maximum 5
   */
  currentStageIdx: number;
  selectedPrompts: string[];
  purposeOptions: string[];
  /** @nullable */
  purposeStatement?: string | null;
  season?: Season;
  actionPlan?: ActionPlan;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  completedAt?: Date | null;
}