# lib/api-zod/src/generated/types/journeyUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionPlan } from "./actionPlan";
import type { JourneyUpdateStatus } from "./journeyUpdateStatus";
import type { Season } from "./season";

export interface JourneyUpdate {
  /**
   * @minimum 0
   * @maximum 5
   */
  currentStageIdx?: number;
  selectedPrompts?: string[];
  purposeOptions?: string[];
  /** @nullable */
  purposeStatement?: string | null;
  status?: JourneyUpdateStatus;
  season?: Season;
  actionPlan?: ActionPlan;
}