# lib/api-zod/src/generated/types/purposeThemeCreate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Evidence } from "./evidence";

export interface PurposeThemeCreate {
  position?: number;
  name: string;
  description: string;
  evidence?: Evidence[];
}