# lib/api-zod/src/generated/types/season.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Season {
  summary?: string;
  roles?: string[];
  capacity?: string;
  constraints?: string[];
  opportunities?: string[];
  expressions?: string[];
}