# lib/api-zod/src/generated/types/storyCardUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoryCardUpdate {
  title?: string;
  summary?: string;
  /** @nullable */
  promptId?: string | null;
  isDifficultExperience?: boolean;
  actions?: string[];
  feelings?: string[];
  people?: string[];
  impact?: string[];
}